Create DesignPattern.c
Sure! Here's a simple C program to print a diamond pattern. You can specify the number of rows for the upper part of the diamond, and the program will print both the upper and lower parts accordingly.

Diamond Pattern in C
c
Copy code
#include <stdio.h>

int main() {
    int n, i, j;

    // Get the number of rows for the upper part of the diamond
    printf("Enter the number of rows: ");
    scanf("%d", &n);

    // Upper part of the diamond
    for (i = 1; i <= n; i++) {
        // Print spaces
        for (j = 1; j <= n - i; j++) {
            printf(" ");
        }
        // Print stars
        for (j = 1; j <= (2 * i - 1); j++) {
            printf("*");
        }
        printf("\n");
    }

    // Lower part of the diamond
    for (i = n - 1; i >= 1; i--) {
        // Print spaces
        for (j = 1; j <= n - i; j++) {
            printf(" ");
        }
        // Print stars
        for (j = 1; j <= (2 * i - 1); j++) {
            printf("*");
        }
        printf("\n");
    }

    return 0;
}
Explanation
Input: The program prompts the user to enter the number of rows for the upper part of the diamond.
Upper Part:
A loop runs from 1 to n to create each row.
For each row, it first prints the required number of spaces, followed by asterisks (*).
The number of asterisks increases as you move down the rows.
Lower Part:
A similar loop runs from n-1 to 1 to create the lower part of the diamond.
Again, spaces are printed first, followed by asterisks, decreasing as you move down.